# contracts/openapi/regional-core-banking-api-v1.yaml
openapi: 3.1.0
info:
  title: La Régionale - Core Banking API
  version: 1.0.0
  summary: Approved Regional-owned Core Banking API V1 contract.
  description: |
    Contract candidate for the autonomous La Régionale Core Banking API.

    This contract is owned by the Regional repository. SIXPAY is one consumer
    among others and its contracts/clients are compatibility evidence only.

    R1 consolidated the operation surfaces already present in the Regional
    draft against available SIXPAY compatibility references and explicit
    Regional decisions. The V1 contract has received explicit human approval.
    No Spring/OpenAPI generation is executed by R1 itself.

  x-regional-contract:
    lifecycleStatus: APPROVED
    approvalStatus: APPROVED
    generationPolicy: ACTIVE
    codeGenerationAllowed: true
    owner: REGIONAL_CORE_BANKING_API
    sourceOfTruth: THIS_REPOSITORY_AFTER_APPROVAL
    sixpayRole: COMPATIBILITY_CONSUMER
    sixpayModelsAreSourceOfTruth: false
    r1Status: APPROVED
    approvalBasis:
      - EXPLICIT_REGIONAL_HUMAN_APPROVAL
      - SIXPAY_WIRE_COMPATIBILITY_EVIDENCE
      - REGIONAL_PAYMENT_EVENT_MAPPING_DECISIONS
      - LEGACY_CUSTOMER_DETAIL_FUNCTIONAL_EVIDENCE
servers:
  - url: https://regional-core-banking-api.example
    description: Environment-specific placeholder. Runtime endpoint is external configuration.
security:
  - oauth2ClientCredentials: []
    mutualTLS: []
...
paths:
  /api/v1/customers:
    get:
      tags: [Customers]
      operationId: searchCustomers
      summary: Search banking customers
      x-regional-review-status: APPROVED
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/FinancialInstitutionCodeQuery'
        - name: niu
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 50 }
        - name: customerNumber
          in: query
          required: false
          schema: { type: string, minLength: 1, maxLength: 50 }
        - name: accountNo
          in: query
          required: false
          description: Legacy-compatible account number filter.
          schema: { type: string, minLength: 1, maxLength: 100 }
        - name: age
          in: query
          required: false
          description: Legacy-compatible agency filter used with accountNo when applicable.
          schema: { type: string, minLength: 1, maxLength: 20 }
        - name: accountTypes
          in: query
          required: false
          description: Legacy-compatible account type filter; may remain unused initially.
          schema: { type: string, minLength: 1, maxLength: 200 }
        - name: chapitre
          in: query
          required: false
          description: Legacy-compatible chapter exclusion filter; may remain unused initially.
          schema: { type: string, minLength: 1, maxLength: 200 }
      responses:
        '200':
          description: Matching customers, possibly empty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerSearchResult' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
        '500': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/customers/{customerReference}:
    get:
      tags: [Customers]
      operationId: getCustomer
      summary: Get banking customer identity/KYC information
      x-regional-review-status: APPROVED
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CustomerReference'
        - $ref: '#/components/parameters/FinancialInstitutionCodeHeader'
      responses:
        '200':
          description: Customer identity found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerIdentity' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
        '500': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/customers/{customerReference}/accounts:
    get:
      tags: [Customers]
      operationId: getCustomerAccounts
      summary: Find accounts for a banking customer
      x-regional-review-status: APPROVED
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CustomerReference'
        - $ref: '#/components/parameters/FinancialInstitutionCodeHeader'
        - name: rib
          in: query
          required: false
          schema: { type: string, minLength: 10, maxLength: 50 }
        - name: iban
          in: query
          required: false
          schema:
            type: string
            minLength: 15
            maxLength: 34
            pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{11,30}$'
      responses:
        '200':
          description: Customer accounts, possibly empty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AccountSearchResult' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '404': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
        '500': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
  /api/v1/customer-verifications:
    post:
      tags: [Customers]
      operationId: verifyCustomer
      summary: Execute a composite banking customer/account verification
      x-regional-review-status: APPROVED
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CustomerVerificationRequest' }
      responses:
        '200':
          description: Completed positive or negative verification result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerVerificationResult' }
        '400': { $ref: '#/components/responses/ProblemResponse' }
        '401': { $ref: '#/components/responses/ProblemResponse' }
        '403': { $ref: '#/components/responses/ProblemResponse' }
        '422': { $ref: '#/components/responses/ProblemResponse' }
        '429': { $ref: '#/components/responses/ProblemResponse' }
        '500': { $ref: '#/components/responses/ProblemResponse' }
        '503': { $ref: '#/components/responses/ProblemResponse' }
...
components:
  securitySchemes:
    oauth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://identity.regional.example/oauth2/token
          scopes: {}
    mutualTLS:
      type: mutualTLS
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      schema: { type: string, format: uuid }
    FinancialInstitutionCodeHeader:
      name: X-Financial-Institution-Code
      in: header
      required: true
      schema: { $ref: '#/components/schemas/FinancialInstitutionCode' }
    FinancialInstitutionCodeQuery:
      name: financialInstitutionCode
      in: query
      required: true
      schema: { $ref: '#/components/schemas/FinancialInstitutionCode' }
...
    CustomerReference:
      name: customerReference
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
...
  schemas:
    FinancialInstitutionCode:
      type: string
      minLength: 1
      maxLength: 50
...
    CustomerSearchResult:
      type: object
      additionalProperties: false
      required: [customers, resultCount]
      properties:
        customers:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/CustomerSummary' }
        resultCount: { type: integer, minimum: 0, maximum: 100 }
    CustomerSummary:
      type: object
      additionalProperties: false
      description: |
        Regional customer summary preserving the existing wire-compatible fields
        while additively retaining the complete supplied legacy CustomerDetail field set.
      required: [customerReference, customerNumber, financialInstitutionCode, niu, legalName]
      properties:
        customerReference: { type: string, minLength: 1, maxLength: 100 }
        customerNumber: { type: string, minLength: 1, maxLength: 50 }
        financialInstitutionCode: { $ref: '#/components/schemas/FinancialInstitutionCode' }
        niu: { type: string, minLength: 1, maxLength: 50 }
        legalName: { type: string, minLength: 1, maxLength: 200 }
        matricule: { type: string, maxLength: 100 }
        accountNo:
          type: string
          maxLength: 100
          description: Legacy account reference represented as `age-ncp-clc`.
        accountName: { type: string, maxLength: 200 }
        customerName: { type: string, maxLength: 200 }
        sexe: { type: string, maxLength: 20 }
        adresse: { type: string, maxLength: 500 }
        dataNaissance:
          type: string
          maxLength: 50
          description: Preserved as String to match the supplied legacy CustomerDetail.
        lieuNaissance: { type: string, maxLength: 200 }
        departNaissance: { type: string, maxLength: 200 }
        situation: { type: string, maxLength: 100 }
        typePiece: { type: string, maxLength: 100 }
        numCNI: { type: string, maxLength: 100 }
        dateDelivrance: { type: string, format: date }
        lieuDelivrance: { type: string, maxLength: 200 }
        dateExpiration: { type: string, format: date }
        ville: { type: string, maxLength: 200 }
        codeGestionnaire: { type: string, maxLength: 100 }
        loginGestionnaire: { type: string, maxLength: 100 }
        comptes:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 100 }
        telephones:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 50 }
        adresseMails:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 254 }
    CustomerIdentity:
      type: object
      additionalProperties: false
      description: |
        Regional customer identity preserving the existing wire-compatible fields
        while additively retaining the complete supplied legacy CustomerDetail field set.
      required:
        [customerReference, customerNumber, financialInstitutionCode, niu, legalName, kycStatus, source, retrievedAt]
      properties:
        customerReference: { type: string, minLength: 1, maxLength: 100 }
        customerNumber: { type: string, minLength: 1, maxLength: 50 }
        financialInstitutionCode: { $ref: '#/components/schemas/FinancialInstitutionCode' }
        niu: { type: string, minLength: 1, maxLength: 50 }
        legalName: { type: string, minLength: 1, maxLength: 200 }
        phoneNumber: { type: [string, 'null'], maxLength: 20 }
        email: { type: [string, 'null'], format: email }
        kycStatus:
          type: string
          enum: [COMPLETE, INCOMPLETE, EXPIRED, NOT_VERIFIED, UNKNOWN]
        source: { type: string, const: AMPLITUDE }
        retrievedAt: { type: string, format: date-time }
        matricule: { type: string, maxLength: 100 }
        accountNo:
          type: string
          maxLength: 100
          description: Legacy account reference represented as `age-ncp-clc`.
        accountName: { type: string, maxLength: 200 }
        customerName: { type: string, maxLength: 200 }
        sexe: { type: string, maxLength: 20 }
        adresse: { type: string, maxLength: 500 }
        dataNaissance:
          type: string
          maxLength: 50
          description: Preserved as String to match the supplied legacy CustomerDetail.
        lieuNaissance: { type: string, maxLength: 200 }
        departNaissance: { type: string, maxLength: 200 }
        situation: { type: string, maxLength: 100 }
        typePiece: { type: string, maxLength: 100 }
        numCNI: { type: string, maxLength: 100 }
        dateDelivrance: { type: string, format: date }
        lieuDelivrance: { type: string, maxLength: 200 }
        dateExpiration: { type: string, format: date }
        ville: { type: string, maxLength: 200 }
        codeGestionnaire: { type: string, maxLength: 100 }
        loginGestionnaire: { type: string, maxLength: 100 }
        comptes:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 100 }
        telephones:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 50 }
        adresseMails:
          type: array
          uniqueItems: true
          items: { type: string, maxLength: 254 }
    AccountSearchResult:
      type: object
      additionalProperties: false
      required: [accounts, resultCount]
      properties:
        accounts:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/AccountSummary' }
        resultCount: { type: integer, minimum: 0, maximum: 100 }
    AccountSummary:
      type: object
      additionalProperties: false
      required: [accountReference, customerReference, status, currency]
      properties:
        accountReference: { type: string, minLength: 1, maxLength: 100 }
        customerReference: { type: string, minLength: 1, maxLength: 100 }
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BLOCKED, CLOSED, UNKNOWN]
        currency: { type: string, minLength: 3, maxLength: 3 }
    CustomerVerificationRequest:
      type: object
      additionalProperties: false
      required: [accountReference, expectedNiu, expectedAccountHolder, requestedAt]
      properties:
        accountReference: { type: string, minLength: 1, maxLength: 100 }
        expectedNiu: { type: string, minLength: 1, maxLength: 50 }
        expectedAccountHolder: { type: string, minLength: 1, maxLength: 200 }
        requiredKycFields:
          type: array
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 64 }
        requestedAt: { type: string, format: date-time }
    VerificationCheck:
      type: object
      additionalProperties: false
      required: [type, result]
      properties:
        type:
          type: string
          enum:
            - CUSTOMER_EXISTS
            - FINANCIAL_INSTITUTION_MATCHES
            - NIU_MATCHES
            - IDENTITY_MATCHES
            - ACCOUNT_EXISTS
            - ACCOUNT_BELONGS_TO_CUSTOMER
            - ACCOUNT_IS_ACTIVE
            - ACCOUNT_NOT_BLOCKED
            - ACCOUNT_NOT_OPPOSED
            - REQUIRED_KYC_PRESENT
            - REQUIRED_KYC_VERIFIED
        result:
          type: string
          enum: [PASS, FAIL, UNKNOWN]
        reasonCode:
          type: [string, 'null']
          maxLength: 50
        checkedAt:
          type: [string, 'null']
          format: date-time
    CustomerVerificationResult:
      type: object
      additionalProperties: false
      required: [verificationId, verifiedAt, source, outcome, checks]
      properties:
        verificationId:
          type: string
          format: uuid
        verifiedAt:
          type: string
          format: date-time
        source:
          type: string
          const: AMPLITUDE
        outcome:
          type: string
          enum: [VERIFIED, REJECTED, INDETERMINATE]
          description: |
            VERIFIED requires every applicable mandatory check to pass.
            REJECTED is a conclusive negative banking result.
            INDETERMINATE never authorizes progression.
        customerReference:
          type: [string, 'null']
          minLength: 1
          maxLength: 100
        accountReference:
          type: [string, 'null']
          minLength: 1
          maxLength: 100
        checks:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/VerificationCheck'
        identity:
          oneOf:
            - $ref: '#/components/schemas/CustomerIdentity'
            - type: 'null'
        account:
          oneOf:
            - $ref: '#/components/schemas/AccountSummary'
            - type: 'null'
...
    Problem:
      type: object
      additionalProperties: true
      required: [status, title]
      properties:
        status: { type: integer, minimum: 400, maximum: 599 }
        title: { type: string }
        detail: { type: [string, 'null'] }
        code: { type: [string, 'null'] }
        correlationId: { type: [string, 'null'] }
  responses:
    ProblemResponse:
      description: Regional Core Banking API error response.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }